refactor: hooking restructure
