+++ BigBaseV2/src/hooks/gui/swap_chain_resizebuffers.cpp
@@ -1,0 +1,32 @@
+#include "hooking.hpp"
+#include "renderer.hpp"
+#include "script.hpp"
+
+namespace big
+{
+	HRESULT hooks::swapchain_resizebuffers(IDXGISwapChain* this_, UINT buffer_count, UINT width, UINT height, DXGI_FORMAT new_format, UINT swapchain_flags)
+	{
+		TRY_CLAUSE
+		{
+			if (g_running)
+			{
+				g_renderer->pre_reset();
+
+				const auto result = g_hooking->m_swapchain_hook.get_original<decltype(&swapchain_resizebuffers)>(swapchain_resizebuffers_index)
+					(this_, buffer_count, width, height, new_format, swapchain_flags);
+
+				if (SUCCEEDED(result))
+				{
+					g_renderer->post_reset();
+				}
+
+				return result;
+			}
+
+			return g_hooking->m_swapchain_hook.get_original<decltype(&swapchain_resizebuffers)>(swapchain_resizebuffers_index)
+				(this_, buffer_count, width, height, new_format, swapchain_flags);
+		} EXCEPT_CLAUSE
+
+		return NULL;
+	}
+}
