+++ BigBaseV2/src/hooks/native_hooks/gta_thread_start.cpp
