+++ BigBaseV2/src/hooks/player_management/network_player_mgr.cpp
