+++ BigBaseV2/src/hooks/player_management/player_left.cpp
