+++ BigBaseV2/src/hooks/protections/increment_stat_event.cpp
