+++ BigBaseV2/src/hooks/protections/net_array_handler.cpp
