+++ BigBaseV2/src/hooks/protections/network_group_override.cpp
