+++ BigBaseV2/src/hooks/protections/received_clone_sync.cpp
