+++ BigBaseV2/src/hooks/protections/received_event.cpp
